reformulate the api call

+++ app/Controller/CollectionsController.php
@@ -24,7 +24,7 @@ class CollectionsController extends AppController {
 										  'User.account_sid' => $credentials['account_sid'])
 				));
                 
-                if (!empty($user) && $this->Auth->login($user)) {
+                if (!empty($user) /*&& $this->Auth->login($user)*/) {
 					$this->Auth->allow($this->params['action']);
 					$this->Session->destroy();
 				}else {
@@ -42,20 +42,39 @@ class CollectionsController extends AppController {
  * @return void
  */
 	public function index() {
-		$this->Collection->recursive = 0;
-		$query = '';
-		if(isset($this->request->query['q'])) {
-			$query = $this->request->query['q'];
-			$this->paginate['conditions'][] = array(
-				'OR' => array(
-					'Collection.title LIKE' => "%$query%",
-					'Collection.subtitle LIKE' => "%$query%",
-					'Collection.description LIKE' => "%$query%"
-				)
-			);			
+		if($this->isRest()){
+			$data = array();
+			$offset = isset($this->request->query['offset']) ? $this->request->query['offset'] : '1';
+			$limit = isset($this->request->query['limit']) ? $this->request->query['limit'] : '10';
+			
+			if(!preg_match('/^[1-9][0-9]*$/',$offset) || !preg_match('/^[1-9][0-9]*$/',$limit)) {
+				$msg = sprintf(__('Please check if "offset" or "limit" are set as numeric numbers from 1 to 999999'));
+				return $this->Rest->abort(array('status' => '400', 'error' => $msg));		
+			}
+			
+			$options = array('limit' => $limit, 'offset'=> $offset - 1 );
+			$data['collections'] = $this->Collection->find('all', $options);
+			$data['total'] = $this->Collection->find('count');
+			$data['offset'] = $offset;
+			$data['limit'] = $limit;
+			
+			$this->set(compact('data'));
+		} else {
+			$this->Collection->recursive = 0;
+			$query = '';
+			if(isset($this->request->query['q'])) {
+				$query = $this->request->query['q'];
+				$this->paginate['conditions'][] = array(
+					'OR' => array(
+						'Collection.title LIKE' => "%$query%",
+						'Collection.subtitle LIKE' => "%$query%",
+						'Collection.description LIKE' => "%$query%"
+					)
+				);
+			}
+			$collections = $this->paginate();
+			$this->set(compact('collections','query'));
 		}
-		$collections = $this->paginate();
-		$this->set(compact('collections','query'));
 	}
 	
 /**
@@ -66,11 +85,20 @@ class CollectionsController extends AppController {
  * @return void
  */
 	public function view($id = null) {
-		if (!$this->Collection->exists($id)) {
-			throw new NotFoundException(__('Invalid collection'));
+		if($this->isRest()){
+			if (!$this->Collection->exists($id)) {
+				return $this->Rest->abort(array('status' => '400', 'error' => __('Invalid collection')));
+			}
+			$options = array('conditions' => array('Collection.' . $this->Collection->primaryKey => $id));
+			$data = $this->Collection->find('first', $options);
+			$this->set(compact('data'));
+		} else {
+			if (!$this->Collection->exists($id)) {
+				throw new NotFoundException(__('Invalid collection'));
+			}
+			$options = array('conditions' => array('Collection.' . $this->Collection->primaryKey => $id));
+			$this->set('collection', $this->Collection->find('first', $options));
 		}
-		$options = array('conditions' => array('Collection.' . $this->Collection->primaryKey => $id));
-		$this->set('collection', $this->Collection->find('first', $options));
 	}
 
 /**
@@ -102,7 +130,7 @@ class CollectionsController extends AppController {
 				'subtitle' => '',
 				'description' => '',
 			);
-			$this->data = array( 'Collection' => $data );
+			$this->data = array('Collection' => $data );
 		}
 	}
 
@@ -114,19 +142,31 @@ class CollectionsController extends AppController {
  * @return void
  */
 	public function edit($id = null) {
-		if (!$this->Collection->exists($id)) {
-			throw new NotFoundException(__('Invalid collection'));
-		}
-		if ($this->request->is('post') || $this->request->is('put')) {
-			if ($this->Collection->save($this->request->data)) {
-				$this->Session->setFlash(__('The collection has been saved'), 'flash/success');
-				$this->redirect(array('action' => 'index'));
+		if($this->isRest()){
+			$data = array();
+			$this->Collection->id = $id;
+			
+			if (!$this->Collection->exists()) {
+				return $this->Rest->abort(array('status' => '400', 'error' => __('Invalid collection')));
 			} else {
-				$this->Session->setFlash(__('The collection could not be saved. Please, try again.'), 'flash/error');
+				
 			}
+			$this->set(compact('data'));		
 		} else {
-			$options = array('conditions' => array('Collection.' . $this->Collection->primaryKey => $id));
-			$this->request->data = $this->Collection->find('first', $options);
+			if (!$this->Collection->exists($id)) {
+				throw new NotFoundException(__('Invalid collection'));
+			}
+			if ($this->request->is('post') || $this->request->is('put')) {
+				if ($this->Collection->save($this->request->data)) {
+					$this->Session->setFlash(__('The collection has been saved'), 'flash/success');
+					$this->redirect(array('action' => 'index'));
+				} else {
+					$this->Session->setFlash(__('The collection could not be saved. Please, try again.'), 'flash/error');
+				}
+			} else {
+				$options = array('conditions' => array('Collection.' . $this->Collection->primaryKey => $id));
+				$this->request->data = $this->Collection->find('first', $options);
+			}
 		}
 	}
 
@@ -139,17 +179,34 @@ class CollectionsController extends AppController {
  * @return void
  */
 	public function delete($id = null) {
-		$this->Collection->id = $id;
-		if (!$this->Collection->exists()) {
-			throw new NotFoundException(__('Invalid collection'));
-		}
-		$this->request->onlyAllow('post', 'delete');
-		if ($this->Collection->delete()) {
-			$this->Session->setFlash(__('Collection deleted'), 'flash/success');
+		if($this->isRest()){
+			$data = array();
+			$this->Collection->id = $id;
+			
+			if (!$this->Collection->exists()) {
+				return $this->Rest->abort(array('status' => '400', 'error' => __('Invalid collection')));
+			} else {
+				if ($this->Collection->delete()) {
+					$data['message'] = __('Collection deleted');
+				} else {
+					return $this->Rest->abort(array('status' => '400', 'error' => __('Fail to delete the collection')));			
+				}
+			}
+			$this->set(compact('data'));
+			
+		} else {
+			$this->Collection->id = $id;
+			if (!$this->Collection->exists()) {
+				throw new NotFoundException(__('Invalid collection'));
+			}
+			$this->request->onlyAllow('post', 'delete');
+			if ($this->Collection->delete()) {
+				$this->Session->setFlash(__('Collection deleted'), 'flash/success');
+				$this->redirect(array('action' => 'index'));
+			}
+			$this->Session->setFlash(__('Collection was not deleted'), 'flash/error');
 			$this->redirect(array('action' => 'index'));
 		}
-		$this->Session->setFlash(__('Collection was not deleted'), 'flash/error');
-		$this->redirect(array('action' => 'index'));
 	}
 /**
  * REST Index method
@@ -187,7 +244,7 @@ class CollectionsController extends AppController {
  */	
 	public function apiv1_view($id = null) {
 		if (!$this->Collection->exists($id)) {
-			$msg = sprintf('Insufficient data');
+			$msg = sprintf('Invalid collection');
 			return $this->Rest->abort(array('status' => '400', 'error' => $msg));
 		}
 		$options = array('conditions' => array('Collection.' . $this->Collection->primaryKey => $id));
